Clean playwright files, Add readme
