--- src/core/tests/auth.json
@@ -1,26 +1,0 @@
-{
-  "cookies": [],
-  "origins": [
-    {
-      "origin": "http://localhost:8081",
-      "localStorage": [
-        {
-          "name": "main",
-          "value": "{\"user\":{\"id\":\"\",\"name\":\"\",\"organization_name\":\"\",\"permissions\":[]},\"vertical_view\":false,\"drawerVisible\":true,\"itemCountTotal\":0,\"itemCountFiltered\":0,\"coreAPIURL\":\"http://127.0.0.1:5000/api\",\"gitInfo\":\"\",\"buildDate\":\"2023-11-18T16:32:09.587Z\",\"sentryDSN\":\"\",\"notification\":{\"message\":\"\",\"type\":\"\",\"show\":false}}"
-        },
-        {
-          "name": "authenticator",
-          "value": "{\"jwt\":\"\",\"login_uri\":\"/login\",\"logout_uri\":\"/logout\",\"external_login_uri\":false,\"external_logout_uri\":false,\"user\":{},\"sub\":\"\",\"exp\":0}"
-        },
-        {
-          "name": "ACCESS_TOKEN",
-          "value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTcwMDMyNTI4OCwianRpIjoiYWMwZGJlMjMtYmZjZi00OTQ4LWI0Y2YtYTUyZWE3OTk2MmFkIiwidHlwZSI6ImFjY2VzcyIsInN1YiI6ImFkbWluIiwibmJmIjoxNzAwMzI1Mjg4LCJleHAiOjE3MDAzMzk2ODgsInVzZXJfY2xhaW1zIjp7ImlkIjoxLCJuYW1lIjoiQXJ0aHVyIERlbnQiLCJvcmdhbml6YXRpb25fbmFtZSI6IlRoZSBFYXJ0aCIsInBlcm1pc3Npb25zIjpbIkNPTkZJR19ST0xFX0RFTEVURSIsIkNPTkZJR19CT1RfVVBEQVRFIiwiQ09ORklHX0JPVF9BQ0NFU1MiLCJBTkFMWVpFX0FDQ0VTUyIsIk1ZX0FTU0VUU19DUkVBVEUiLCJDT05GSUdfV09SRF9MSVNUX0FDQ0VTUyIsIkNPTkZJR19BQ0xfVVBEQVRFIiwiQ09ORklHX0FUVFJJQlVURV9BQ0NFU1MiLCJDT05GSUdfUFVCTElTSEVSX1BSRVNFVF9DUkVBVEUiLCJDT05GSUdfUFVCTElTSEVSX1BSRVNFVF9ERUxFVEUiLCJBTkFMWVpFX1VQREFURSIsIkNPTkZJR19VU0VSX0NSRUFURSIsIkFTU0VTU19DUkVBVEUiLCJDT05GSUdfUFJPRFVDVF9UWVBFX0NSRUFURSIsIkNPTkZJR19ST0xFX0NSRUFURSIsIlBVQkxJU0hfVVBEQVRFIiwiQ09ORklHX09SR0FOSVpBVElPTl9VUERBVEUiLCJDT05GSUdfUk9MRV9VUERBVEUiLCJBTkFMWVpFX0RFTEVURSIsIkNPTkZJR19XT1JEX0xJU1RfREVMRVRFIiwiQ09ORklHX1VTRVJfQUNDRVNTIiwiQ09ORklHX09SR0FOSVpBVElPTl9DUkVBVEUiLCJDT05GSUdfUFJPRFVDVF9UWVBFX1VQREFURSIsIkNPTkZJR19BUElfQUNDRVNTIiwiTVlfQVNTRVRTX0NPTkZJRyIsIkNPTkZJR19PU0lOVF9TT1VSQ0VfQ1JFQVRFIiwiUFVCTElTSF9QUk9EVUNUIiwiQ09ORklHX1JFUE9SVF9UWVBFX1VQREFURSIsIkFTU0VTU19BQ0NFU1MiLCJDT05GSUdfQk9UX0NSRUFURSIsIkNPTkZJR19PU0lOVF9TT1VSQ0VfR1JPVVBfQUNDRVNTIiwiQ09ORklHX09TSU5UX1NPVVJDRV9HUk9VUF9ERUxFVEUiLCJDT05GSUdfQUNMX0NSRUFURSIsIkNPTkZJR19PU0lOVF9TT1VSQ0VfREVMRVRFIiwiQ09ORklHX1dPUkRfTElTVF9VUERBVEUiLCJDT05GSUdfUkVQT1JUX1RZUEVfQ1JFQVRFIiwiQ09ORklHX1BVQkxJU0hFUl9QUkVTRVRfVVBEQVRFIiwiQ09ORklHX1JFUE9SVF9UWVBFX0FDQ0VTUyIsIkNPTkZJR19PU0lOVF9TT1VSQ0VfQUNDRVNTIiwiQ09ORklHX1VTRVJfREVMRVRFIiwiQ09ORklHX0JPVF9ERUxFVEUiLCJDT05GSUdfVVNFUl9VUERBVEUiLCJDT05GSUdfQUNDRVNTIiwiQ09ORklHX09TSU5UX1NPVVJDRV9HUk9VUF9DUkVBVEUiLCJDT05GSUdfQUNMX0FDQ0VTUyIsIkNPTkZJR19BVFRSSUJVVEVfVVBEQVRFIiwiQ09ORklHX1BVQkxJU0hFUl9QUkVTRVRfQUNDRVNTIiwiQ09ORklHX1BST0RVQ1RfVFlQRV9BQ0NFU1MiLCJDT05GSUdfUkVQT1JUX1RZUEVfREVMRVRFIiwiUFVCTElTSF9ERUxFVEUiLCJDT05GSUdfT1NJTlRfU09VUkNFX1VQREFURSIsIkNPTkZJR19PUkdBTklaQVRJT05fREVMRVRFIiwiUFVCTElTSF9DUkVBVEUiLCJQVUJMSVNIX0FDQ0VTUyIsIkNPTkZJR19XT1JEX0xJU1RfQ1JFQVRFIiwiTVlfQVNTRVRTX0FDQ0VTUyIsIkNPTkZJR19PU0lOVF9TT1VSQ0VfR1JPVVBfVVBEQVRFIiwiQ09ORklHX1BST0RVQ1RfVFlQRV9ERUxFVEUiLCJBU1NFU1NfVVBEQVRFIiwiQ09ORklHX0FDTF9ERUxFVEUiLCJDT05GSUdfT1JHQU5JWkFUSU9OX0FDQ0VTUyIsIkFOQUxZWkVfQ1JFQVRFIiwiQ09ORklHX0FUVFJJQlVURV9DUkVBVEUiLCJDT05GSUdfV09SS0VSX0FDQ0VTUyIsIkNPTkZJR19ST0xFX0FDQ0VTUyIsIkFTU0VTU19ERUxFVEUiLCJDT05GSUdfQVRUUklCVVRFX0RFTEVURSJdfX0.NW2sB5FlhWIYjwJSQgaA8N90tbeyT8koQAQnXyqvxko"
-        },
-        {
-          "name": "assess",
-          "value": "{\"osint_sources\":{\"total_count\":0,\"items\":[]},\"osint_source_groups\":{\"total_count\":0,\"items\":[]}}"
-        }
-      ]
-    }
-  ]
-}
